grade-analysis-web/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
pandas==2.2.3
openpyxl==3.1.5
google-generativeai==0.8.3
scikit-learn==1.6.0
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0
aiofiles==24.1.0
